Move all types related to executor into `services/executor.rs`
