+++ crates/types/src/services.rs
@@ -1,3 +1,4 @@
 //! Types for specific services
 
+pub mod executor;
 pub mod txpool;
